requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
docstring_parser==0.16
fastapi==0.109.2
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-aiplatform==1.88.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.10.0
google-generativeai==0.3.2
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
numpy==2.2.4
packaging==24.2
proto-plus==1.26.1
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.27.1
websockets==15.0.1
